refactor: refactor current common code
