+++ shared/src/commonMain/kotlin/com/myapplication/common/birds/datasource/data/BirdsRemoteDataSource.kt
@@ -2,6 +2,6 @@ package com.myapplication.common.birds.datasource.data
 
 import com.myapplication.common.birds.datasource.model.BirdImage
 
-interface BirdsRemoteDataSource {
+internal interface BirdsRemoteDataSource {
     suspend fun getBirdImages(): List<BirdImage>
 }
