+++ shared/src/commonMain/kotlin/com/myapplication/common/birds/datasource/data/BirdsRemoteDataSourceImpl.kt
@@ -6,7 +6,7 @@ import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.request.get
 
-class BirdsRemoteDataSourceImpl(
+internal class BirdsRemoteDataSourceImpl(
     private val httpClient: HttpClient = KtorHttpClient(),
 ) : BirdsRemoteDataSource {
