+++ shared/src/commonMain/kotlin/com/myapplication/common/birds/datasource/model/BirdImage.kt
@@ -3,7 +3,7 @@ package com.myapplication.common.birds.datasource.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class BirdImage(
+internal data class BirdImage(
     val author: String,
     val category: String,
     val path: String,
