--- shared/src/commonMain/kotlin/com/myapplication/common/birds/datasource/model/BirdsError.kt
@@ -1,6 +1,0 @@
-package com.myapplication.common.birds.datasource.model
-
-sealed class BirdsError {
-    object Network : BirdsError()
-    object Unknown : BirdsError()
-}
