+++ shared/src/commonMain/kotlin/com/myapplication/common/birds/datasource/model/GetBirdsError.kt
@@ -1,0 +1,6 @@
+package com.myapplication.common.birds.datasource.model
+
+internal sealed class GetBirdsError {
+    object Network : GetBirdsError()
+    object Unknown : GetBirdsError()
+}
