+++ shared/src/commonMain/kotlin/com/myapplication/common/birds/ui/viewmodel/BirdsViewModel.kt
@@ -1,33 +1,18 @@
 package com.myapplication.common.birds.ui.viewmodel
 
 import arrow.core.Either
-import com.myapplication.common.birds.datasource.model.BirdsError
+import com.myapplication.common.birds.datasource.model.GetBirdsError
 import com.myapplication.common.birds.datasource.repository.BirdsRepository
 import com.myapplication.common.birds.datasource.repository.BirdsRepositoryImpl
-import com.myapplication.common.birds.ui.model.Bird
+import com.myapplication.common.birds.ui.contract.BirdsUiState
+import com.myapplication.common.birds.ui.contract.model.Bird
 import dev.icerock.moko.mvvm.viewmodel.ViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
-sealed interface BirdsUiState {
-    object Loading : BirdsUiState
-    object Failure : BirdsUiState
-    data class Success(
-        val images: List<Bird>,
-        val selectedCategory: String?,
-    ) : BirdsUiState {
-        val categories: Set<String> = images.map { it.category }.toSet()
-        val selectedImages: List<Bird> = images.filter { it.category == selectedCategory }
-    }
-
-    companion object {
-        val initialState = Loading
-    }
-}
-
-class BirdsViewModel(
+internal class BirdsViewModel(
     private val birdsRepository: BirdsRepository = BirdsRepositoryImpl(),
     initialState: BirdsUiState = BirdsUiState.initialState,
 ) : ViewModel() {
@@ -71,5 +56,5 @@ class BirdsViewModel(
         }
     }
 
-    private suspend fun getImages(): Either<BirdsError, List<Bird>> = birdsRepository.getBirds()
+    private suspend fun getImages(): Either<GetBirdsError, List<Bird>> = birdsRepository.getBirds()
 }
